requirements.txt
numpy==1.24.3
opencv_python==4.7.0.72
Pillow==10.1.0
PyYAML==6.0.1
Requests==2.31.0
scipy==1.9.1
scikit-image==0.22.0
timm==0.9.2
torch>=1.7
torchvision
tqdm==4.65.0
wandb==0.15.5
tensorboard
matplotlib
lmdb==1.4.1
dlib==19.24.2
huggingface_hub
